Makefile: set CSV 'replaces' on prepare-release
This updates and relocates the 'sed' command updating the 'replaces' key
on the CSV file.

For some reason, the variable PREVIOUS_VERSION is updated after
bundle is updated with the next version's number.
After the 'bundle' target is executed the value of PREVIOUS_VERSION
during execution of 'bundle-image' target is updated because apparently
'utils/get-current-version.sh' is executed again.