Add value selection to variables
The Variables in CO have predefined selections that are parsed from the
Data Stream.
These predefined selections are used to customize a rule to a specific
policy or common use cases.

CO doesn't take advantage of these selector.
This patch allows the user to tailor a variable using the available
selections in the variable.